requirements.txt
#
# Use this file to add Python dependencies.
#
# Install dependencies with:
#    python3 -m pip install -r requirements.txt
# Update them with:
#    python3 -m pip install -U -r requirement.txt
#

# GObject
# pygobject == 3.27.1

# SQLAlchemy ORM
sqlalchemy >= 1.2.7
sqlalchemy-utils >= 0.32.21
alembic >= 0.9.9

# BDD / Tests
behave >= 1.2.5

# Coding standards
flake8 >= 3.5.0
flake8-docstrings >= 1.3.0

# Python
Pillow >= 3.2.0
python-dateutil >= 2.7.3
